♻️ refactor: satisfy SwiftLint length/closure caps after card conversion
Split ScenarioDetailView section builders into a sibling-file extension
(type_body_length); extract scenariosCard / Gallery header+details+footer
helpers (function_body_length); hoist enumerated() collections to locals so
ForEach closure params stay on the brace line (closure_parameter_position).
No behavior change.
